Updated the server to work with etcd
The earlier versions of this software required regular polling by the
viewer applications in order to detect if any settings on the server had
changed. Now the viewer apps watch for their prefix on etcd, and when a
change is made on the server, that change is written to etcd. The viewer
app then sees that change and reacts.